Pass success/failure through IScmBackup decorators
We want to be able to return a process exit code that indicates success
or failure of the backup. Currently, exceptions are caught and logged at
the ErrorHandlingScmBackup layer of the decorator chain, which means
that the process entry point is unaware that something has gone wrong.

To solve this, we now pass a Boolean true/false value up through the
decorator chain. This isn't 100% clear in all cases, because most errors
will bubble up via exceptions rather than this return value. However,
perfection is the enemy of the good, and this approach is still useful.