remove call to workspaces API and directly call repositories API
(fixes #50)

Call to workspaces API is not necessary if all we need is the repositories API URL.
The repositories API accepts a workspace ID anyway.
(Bitbucket support, Ticket BBS-134983)

Note: the parameter must be the workspace ID, no matter which of the two APIs are called.
There's no difference between users and teams anymore!